refactor: rename module after sole exported function
